+++ src/impl/xliff-message-parser.spec.ts
@@ -1,6 +1,7 @@
 import {XliffMessageParser} from './xliff-message-parser';
 import {ParsedMessage} from './parsed-message';
 import {DOMParser} from 'xmldom';
+import {INormalizedMessage} from '../api';
 /**
  * Created by martin on 17.05.2017.
  * Testcases for parsing normalized messages to XLIFF 1.2 and vive versa.
@@ -69,7 +70,7 @@ describe('message parse XLIFF 1.2 test spec', () => {
             let normalizedMessage = 'a text <b>with</b> a bold text';
             let parsedMessage = parsedMessageFor(normalizedMessage);
             expect(parsedMessage.asDisplayString()).toBe(normalizedMessage);
-            expect(parsedMessage.asNativeString()).toBe('a text <x id="START_BOLD_TEXT" ctype="x-b"/>with<x id="CLOSE_BOLD_TEXT" ctype="x-b"/> a bold text');
+            expect(parsedMessage.asNativeString()).toBe('a text <x id="START_BOLD_TEXT" ctype="x-b" equiv-text="&lt;b>"/>with<x id="CLOSE_BOLD_TEXT" ctype="x-b"/> a bold text');
             checkToXmlAndBack(normalizedMessage);
         });
 
@@ -77,7 +78,7 @@ describe('message parse XLIFF 1.2 test spec', () => {
             let normalizedMessage = 'a text <i>with</i> emphasis';
             let parsedMessage = parsedMessageFor(normalizedMessage);
             expect(parsedMessage.asDisplayString()).toBe(normalizedMessage);
-            expect(parsedMessage.asNativeString()).toBe('a text <x id="START_ITALIC_TEXT" ctype="x-i"/>with<x id="CLOSE_ITALIC_TEXT" ctype="x-i"/> emphasis');
+            expect(parsedMessage.asNativeString()).toBe('a text <x id="START_ITALIC_TEXT" ctype="x-i" equiv-text="&lt;i>"/>with<x id="CLOSE_ITALIC_TEXT" ctype="x-i"/> emphasis');
             checkToXmlAndBack(normalizedMessage);
         });
 
@@ -85,7 +86,7 @@ describe('message parse XLIFF 1.2 test spec', () => {
             let normalizedMessage = 'a text with <strange>strange emphasis</strange>';
             let parsedMessage = parsedMessageFor(normalizedMessage);
             expect(parsedMessage.asDisplayString()).toBe(normalizedMessage);
-            expect(parsedMessage.asNativeString()).toBe('a text with <x id="START_TAG_STRANGE" ctype="x-strange"/>strange emphasis<x id="CLOSE_TAG_STRANGE" ctype="x-strange"/>');
+            expect(parsedMessage.asNativeString()).toBe('a text with <x id="START_TAG_STRANGE" ctype="x-strange" equiv-text="&lt;strange>"/>strange emphasis<x id="CLOSE_TAG_STRANGE" ctype="x-strange"/>');
             checkToXmlAndBack(normalizedMessage);
         });
 
@@ -93,15 +94,15 @@ describe('message parse XLIFF 1.2 test spec', () => {
             let normalizedMessage = '<b><i><strange>Placeholder {{0}}</strange></i></b>';
             let parsedMessage = parsedMessageFor(normalizedMessage);
             expect(parsedMessage.asDisplayString()).toBe(normalizedMessage);
-            expect(parsedMessage.asNativeString()).toBe('<x id="START_BOLD_TEXT" ctype="x-b"/><x id="START_ITALIC_TEXT" ctype="x-i"/><x id="START_TAG_STRANGE" ctype="x-strange"/>Placeholder <x id="INTERPOLATION"/><x id="CLOSE_TAG_STRANGE" ctype="x-strange"/><x id="CLOSE_ITALIC_TEXT" ctype="x-i"/><x id="CLOSE_BOLD_TEXT" ctype="x-b"/>');
+            expect(parsedMessage.asNativeString()).toBe('<x id="START_BOLD_TEXT" ctype="x-b" equiv-text="&lt;b>"/><x id="START_ITALIC_TEXT" ctype="x-i" equiv-text="&lt;i>"/><x id="START_TAG_STRANGE" ctype="x-strange" equiv-text="&lt;strange>"/>Placeholder <x id="INTERPOLATION"/><x id="CLOSE_TAG_STRANGE" ctype="x-strange"/><x id="CLOSE_ITALIC_TEXT" ctype="x-i"/><x id="CLOSE_BOLD_TEXT" ctype="x-b"/>');
             checkToXmlAndBack(normalizedMessage);
         });
 
-        it('should parse empty tag like <br/>', () => {
-            let normalizedMessage = 'one line<br/>second line';
+        it('should parse empty tag like <br>', () => {
+            let normalizedMessage = 'one line<br>second line';
             let parsedMessage = parsedMessageFor(normalizedMessage);
             expect(parsedMessage.asDisplayString()).toBe(normalizedMessage);
-            expect(parsedMessage.asNativeString()).toBe('one line<x id="LINE_BREAK" ctype="lb"/>second line');
+            expect(parsedMessage.asNativeString()).toBe('one line<x id="LINE_BREAK" ctype="lb" equiv-text="&lt;br/>"/>second line');
             checkToXmlAndBack(normalizedMessage);
         });
 
@@ -154,6 +155,32 @@ describe('message parse XLIFF 1.2 test spec', () => {
         it('should parse message with 2 embedded ICU message reference', () => {
             let parsedMessage = parsedMessageFromXML('first: <x id="ICU"/>, second <x id="ICU_1"/>');
             expect(parsedMessage.asDisplayString()).toBe('first: <ICU-Message-Ref_0/>, second <ICU-Message-Ref_1/>');
+        });
+
+        it('should set correct placeholer index (issue #84 ngx-i18nsupport) ', () => {
+            const messageWith2Indexes = 'New <x id="START_TAG_XY" ctype="x-xy" equiv-text="&lt;xy&gt;"/><x id="CLOSE_TAG_XY" ctype="x-xy" equiv-text="&lt;/xy&gt;"/> was reported by <x id="START_TAG_XY_1" ctype="x-xy" equiv-text="&lt;xy&gt;"/><x id="CLOSE_TAG_XY" ctype="x-xy" equiv-text="&lt;/xy&gt;"/>';
+            let parsedMessage = parsedMessageFromXML(messageWith2Indexes);
+            const normalizedMessageString = parsedMessage.asDisplayString();
+            expect(normalizedMessageString).toBe('New <xy></xy> was reported by <xy id="1"></xy>');
+            const translatedMessage: INormalizedMessage = parsedMessage.translate('New <xy></xy> was reported by <xy id="1"></xy>');
+            expect(translatedMessage.asNativeString()).toBe('New <x id="START_TAG_XY" ctype="x-xy" equiv-text="&lt;xy>"/><x id="CLOSE_TAG_XY" ctype="x-xy"/> was reported by <x id="START_TAG_XY_1" ctype="x-xy" equiv-text="&lt;xy>"/><x id="CLOSE_TAG_XY" ctype="x-xy"/>');
+        });
+
+        it('should parse simple plural ICU message', () => {
+            let parsedMessage = parsedMessageFromXML('{VAR_PLURAL, plural, =0 {just now} =1 {one minute ago} other {a few minutes ago} }');
+            expect(parsedMessage.asDisplayString()).toBe('<ICU-Message/>');
+            expect(parsedMessage.getICUMessage()).toBeTruthy();
+            const icuMessage = parsedMessage.getICUMessage();
+            expect(icuMessage.getCategories().length).toBe(3);
+        });
+
+        it('should parse plural ICU message with placeholder', () => {
+            let parsedMessage = parsedMessageFromXML('{VAR_PLURAL, plural, =0 {just now} =1 {one minute ago} other {<x id="INTERPOLATION" equiv-text="{{minutes}}"/> minutes ago} }');
+            expect(parsedMessage.asDisplayString()).toBe('<ICU-Message/>');
+            expect(parsedMessage.getICUMessage()).toBeTruthy();
+            const icuMessage = parsedMessage.getICUMessage();
+            expect(icuMessage.getCategories().length).toBe(3);
+            expect(icuMessage.getCategories()[2].getMessageNormalized().asDisplayString()).toBe('{{0}} minutes ago');
         });
 
     });